Controller Server/requirements.txt
websockets
qrcode
pillow
 

